Move code from messages/ __init__.py to utils.py
