+++ kitsune/messages/tests/test_context_processors.py
@@ -4,7 +4,7 @@ from django.test.client import RequestFactory
 import mock
 from nose.tools import eq_
 
-from kitsune import messages
+from kitsune.messages import context_processors
 from kitsune.messages.context_processors import unread_message_count
 from kitsune.sumo.tests import TestCase
 from kitsune.users.tests import user
@@ -13,9 +13,9 @@ from kitsune.users.tests import user
 class UnreadCountTests(TestCase):
     """Tests for unread_message_count."""
 
-    @mock.patch.object(messages, 'unread_count_for')
+    @mock.patch.object(context_processors, 'unread_count_for')
     def test_anonymous(self, unread_count_for):
-        """Test anonymous user with flag active."""
+        """Test anonymous user."""
         unread_count_for.return_value = 3
         rf = RequestFactory()
         request = rf.get('/')
@@ -23,9 +23,9 @@ class UnreadCountTests(TestCase):
         eq_(0, unread_message_count(request)['unread_message_count'])
         assert not unread_count_for.called
 
-    @mock.patch.object(messages, 'unread_count_for')
+    @mock.patch.object(context_processors, 'unread_count_for')
     def test_authenticated(self, unread_count_for):
-        """Test authenticated user with flag active."""
+        """Test authenticated user."""
         unread_count_for.return_value = 3
         rf = RequestFactory()
         request = rf.get('/')
